refactor: replace renderLogout function with class Header
