+++ src/css/helpers/variables.scss
@@ -3,8 +3,8 @@ $colorBlack: rgb(0, 0, 0);
 $colorBlackTransparent: rgba(0, 0, 0, 0.5);
 $colorBlue: rgb(0, 0, 255);
 $colorRed: rgb(255, 0, 0);
-$colorViolet: #9c27b0;
-$colorHover: #ff9640;
+$colorViolet: rgb(156, 39, 176);
+$colorHover: rgb(255, 150, 64);
 
 $space-xxs: 10px;
 $space-xs: 20px;
