+++ src/html/index.html
@@ -18,7 +18,7 @@
       </div>
     </div>
     <div id="root">
-      <div id="logout"></div>
+      <div id="header"></div>
       <div id="main"></div>
       <div id="sidebar"></div>
     </div>
