+++ src/scripts/App.js
@@ -1,16 +1,21 @@
-import { SIDEBAR, MAIN, LOGOUT } from './helpers/variables';
+import { SIDEBAR, MAIN, HEADER } from './helpers/variables';
+import { store } from '../store/index';
 import { createSidebar } from './burger';
-import { createLogout } from './logout';
+import { Header } from './Header';
 import { Authorization } from './Authorization';
+import { router } from '../routes/index';
 
 export class App {
   static reRender(page) {
-    if (page === 'login' || page === 'registration') {
+    const isAuthPage = ['login', 'registration'].includes(page);
+    const isAuthorized = !!store.user.auth.token;
+    if (isAuthPage) {
+      if (isAuthorized) return router.navigate('/');
       SIDEBAR.innerHTML = '';
-      LOGOUT.innerHTML = '';
+      HEADER.innerHTML = '';
     } else {
       this.checkSideBar();
-      this.checkLogout();
+      this.checkHeader();
     }
     MAIN.innerHTML = '';
     this.setContent(page);
@@ -22,9 +27,9 @@ export class App {
     }
   }
 
-  static checkLogout() {
-    if (!LOGOUT.innerHTML) {
-      createLogout();
+  static checkHeader() {
+    if (!HEADER.innerHTML) {
+      Header.render();
     }
   }
